fix various -Wformat

+++ libs/backends/coreaudio/coreaudio_backend.cc
@@ -927,7 +927,7 @@ CoreAudioBackend::coremidi_rediscover()
 
 	for (std::vector<CoreBackendPort*>::iterator it = _system_midi_out.begin (); it != _system_midi_out.end ();) {
 		bool found = false;
-		for (int i = 0; i < _midiio->n_midi_outputs(); ++i) {
+		for (size_t i = 0; i < _midiio->n_midi_outputs(); ++i) {
 			if ((*it)->name() == _midiio->port_name(i, false)) {
 				found = true;
 				break;
@@ -947,7 +947,7 @@ CoreAudioBackend::coremidi_rediscover()
 
 	for (std::vector<CoreBackendPort*>::iterator it = _system_midi_in.begin (); it != _system_midi_in.end ();) {
 		bool found = false;
-		for (int i = 0; i < _midiio->n_midi_inputs(); ++i) {
+		for (size_t i = 0; i < _midiio->n_midi_inputs(); ++i) {
 			if ((*it)->name() == _midiio->port_name(i, true)) {
 				found = true;
 				break;
@@ -965,7 +965,7 @@ CoreAudioBackend::coremidi_rediscover()
 		}
 	}
 
-	for (int i = 0; i < _midiio->n_midi_inputs(); ++i) {
+	for (size_t i = 0; i < _midiio->n_midi_inputs(); ++i) {
 		std::string name = _midiio->port_name(i, true);
 		if (find_port_in(_system_midi_in, name)) {
 			continue;
@@ -986,7 +986,7 @@ CoreAudioBackend::coremidi_rediscover()
 		_port_change_flag = true;
 	}
 
-	for (int i = 0; i < _midiio->n_midi_outputs(); ++i) {
+	for (size_t i = 0; i < _midiio->n_midi_outputs(); ++i) {
 		std::string name = _midiio->port_name(i, false);
 		if (find_port_in(_system_midi_out, name)) {
 			continue;
